requirements.txt
pytenable
pandas
openpyxl
argparse
python-dotenv
coloredlogs
colorama
unidecode